requirements.txt
dill==0.2.6
Django==1.8.4
enum34==1.1.6
future==0.16.0
futures==3.0.5
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
google-gax==0.15.8
googleapis-common-protos==1.5.2
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.1.3
httplib2==0.10.3
jws==0.1.3
MySQL-python==1.2.5
oauth2client==3.0.0
pbr==1.8.0
ply==3.8
protobuf==3.2.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-firebase==1.2
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.9.0
stevedore==1.8.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
